feat(tiff): add LONG8/IFD/IFD8 data types and Values64 accessor
Widens Entry.Count to uint64 and valueBytes to [8]byte to accommodate
BigTIFF entries, with an inlineCap field that preserves the 4-byte
classic-TIFF inline semantics when the IFD walker leaves it unset.

Also adds minimal uint64 casts in ifd.go readEntry to compile under
the widened field types (source-compatible, no behavior change).

+++ internal/tiff/ifd.go
@@ -78,8 +78,8 @@ func readEntry(b *byteReader, offset int64) (Entry, error) {
 	var e Entry
 	e.Tag = tag
 	e.Type = DataType(typ)
-	e.Count = count
-	e.valueOrOffset = vo
+	e.Count = uint64(count)
+	e.valueOrOffset = uint64(vo)
 	copy(e.valueBytes[:], cell)
 	return e, nil
 }
